module_5/requirements.txt

flask
psycopg[binary]
pytest
pytest-cov
beautifulsoup4
urllib3
python-dotenv

